Verificacao de Permissoes

+++ app/Http/Controllers/ProjectController.php
@@ -5,7 +5,8 @@ namespace CodeProject\Http\Controllers;
 use Illuminate\Http\Request;
 use CodeProject\Repositories\ProjectRepository;
 use CodeProject\Services\ProjectService;
-
+use CodeProject\Repositories\ProjectRepositoryEloquent;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 
 class ProjectController extends Controller
 {
@@ -25,7 +26,7 @@ class ProjectController extends Controller
      */
     public function index(ProjectRepository $repository)
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id'=>Authorizer::getResourceOwnerId()]);
     }
 
    
@@ -37,6 +38,7 @@ class ProjectController extends Controller
      */
     public function store(Request $request)
     {
+
         return $this->service->create($request->all());
     }
 
@@ -48,6 +50,9 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+        if($this->checkProjectPermissions($id)==false){
+            return ['error' => 'Access Forbidden'];
+        }
         return $this->repository->find($id);
     }
 
@@ -60,6 +65,9 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id, $noteId)
     {
+       if($this->checkProjectOwner($id)==false){
+            return ['error' => 'Access Forbidden'];
+        } 
        return $this->repository->find($id)->update($request->all());
     }
 
@@ -71,6 +79,31 @@ class ProjectController extends Controller
      */
     public function destroy($id, $noteId)
     {
+        if($this->checkProjectOwner($id)==false){
+            return ['error' => 'Access Forbidden'];
+        }
+
         return $this->repository->delete($noteId);
+    }
+
+    private function checkProjectOwner($projectId){
+
+        $userId = Authorizer::getResourceOwnerId();
+        
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+    private function checkProjectMember($projectId){
+
+        $userId = Authorizer::getResourceOwnerId();
+        
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    private function checkProjectPermissions($projectId){
+        if($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)){
+            return true;
+        }
+        return false;
     }
 }
